+++ backend/src/main/java/com/ll/cafeservice/global/security/exception/InvalidCredentialsException.java
@@ -1,4 +1,4 @@
-package com.ll.cafeservice.security.exception;
+package com.ll.cafeservice.global.security.exception;
 
 public class InvalidCredentialsException extends RuntimeException {
     public InvalidCredentialsException(String message) {
