+++ backend/src/main/java/com/ll/cafeservice/global/security/jwt/JwtProvider.java
@@ -1,6 +1,6 @@
-package com.ll.cafeservice.security.jwt;
+package com.ll.cafeservice.global.security.jwt;
 
-import com.ll.cafeservice.security.exception.CustomJwtException;
+import com.ll.cafeservice.global.security.exception.CustomJwtException;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
